backend/app/integrations/webdav.py
"""
FluxFile - WebDAV 集成模块
============================

通过 WsgiDAV 库提供 WebDAV 协议支持，允许客户端通过标准文件管理器
（如 Windows 资源管理器、macOS Finder）直接访问 FluxFile。

关键设计：
1. 使用 WSGIMiddleware 将 WsgiDAV 挂载到 FastAPI
2. 共享 FastAPI 的认证上下文
3. 与 FluxFile 的权限系统集成

挂载路径：/webdav
"""

import os
from pathlib import Path
from typing import Any, Callable, Dict, List, Optional

from app.core.config import settings
from app.core.logging import get_logger

logger = get_logger(__name__)


# ============================================================================
# WebDAV 认证域提供器
# ============================================================================

class FluxFileDomainController:
    """
    FluxFile 域控制器
    
    为 WsgiDAV 提供认证和授权功能。
    与 FastAPI 的认证系统共享用户验证逻辑。
    """
    
    def __init__(self, realm: str = "FluxFile WebDAV"):
        self.realm = realm
    
    def get_domain_realm(
        self,
        path: str,
        environ: Dict[str, Any],
    ) -> str:
        """返回认证域名称"""
        return self.realm
    
    def require_authentication(
        self,
        realm: str,
        environ: Dict[str, Any],
    ) -> bool:
        """是否需要认证"""
        # 可以根据路径决定是否需要认证
        return True
    
    def is_share_anonymous(
        self,
        path: str,
    ) -> bool:
        """是否允许匿名访问"""
        return False
    
    def basic_auth_user(
        self,
        realm: str,
        username: str,
        password: str,
        environ: Dict[str, Any],
    ) -> bool:
        """
        Basic Auth 用户验证
        
        这里应该调用 FluxFile 的认证系统进行验证。
        可以支持：
        - 本地用户数据库
        - LDAP 认证
        - JWT Token（通过特殊处理）
        """
        # TODO: 集成实际的认证逻辑
        if settings.DEBUG:
            # 开发模式下允许任意用户
            logger.warning(f"WebDAV 开发模式: 允许用户 {username} 访问")
            return True
        
        # 实际应用中应验证用户
        return self._verify_user(username, password)
    
    def _verify_user(self, username: str, password: str) -> bool:
        """实际的用户验证逻辑"""
        # TODO: 实现实际的用户验证
        # 1. 从数据库查询用户
        # 2. 验证密码 hash
        # 3. 或者调用 LDAP 验证
        return False
    
    def supports_http_digest_auth(self) -> bool:
        """是否支持 HTTP Digest 认证"""
        return False


# ============================================================================
# WebDAV 文件系统提供器
# ============================================================================

class FluxFileDAVProvider:
    """
    FluxFile DAV 提供器
    
    将 FluxFile 的文件系统暴露给 WsgiDAV。
    继承 WsgiDAV 的 FilesystemProvider 并添加权限控制。
    """
    
    def __init__(
        self,
        root_path: str,
        readonly: bool = False,
    ):
        self.root_path = Path(root_path).resolve()
        self.readonly = readonly
    
    def get_resource_inst(
        self,
        path: str,
        environ: Dict[str, Any],
    ):
        """
        获取资源实例
        
        这是 WsgiDAV 调用的主要入口。
        """
        # 使用 WsgiDAV 的标准文件系统提供器
        # 添加额外的权限检查
        from wsgidav.fs_dav_provider import FilesystemProvider
        
        provider = FilesystemProvider(str(self.root_path), readonly=self.readonly)
        return provider.get_resource_inst(path, environ)


# ============================================================================
# WebDAV 中间件包装器
# ============================================================================

class FluxFileMiddleware:
    """
    FluxFile 中间件
    
    在 WebDAV 请求前后添加 FluxFile 特定的处理逻辑：
    1. 请求日志记录
    2. 操作审计
    3. 权限检查
    """
    
    def __init__(self, app: Callable):
        self.app = app
    
    def __call__(
        self,
        environ: Dict[str, Any],
        start_response: Callable,
    ):
        # 请求前处理
        method = environ.get("REQUEST_METHOD", "?")
        path = environ.get("PATH_INFO", "/")
        user = environ.get("wsgidav.user", "anonymous")
        
        logger.info(f"WebDAV {method} {path} (用户: {user})")
        
        # 调用 WsgiDAV 应用
        response = self.app(environ, start_response)
        
        # 请求后处理（审计日志等）
        # TODO: 记录到 ClickHouse
        
        return response


# ============================================================================
# 创建 WebDAV 应用
# ============================================================================

def create_webdav_app():
    """
    创建 WsgiDAV 应用实例
    
    配置说明：
    - root_path: 文件系统根目录
    - auth: 认证配置
    - middleware_stack: 中间件栈
    """
    try:
        from wsgidav.wsgidav_app import WsgiDAVApp
        from wsgidav.fs_dav_provider import FilesystemProvider
    except ImportError:
        logger.error(
            "WsgiDAV 未安装，WebDAV 功能不可用。"
            "安装命令: pip install wsgidav cheroot"
        )
        return None
    
    # 配置
    config = {
        # 提供器配置
        "provider_mapping": {
            "/": FilesystemProvider(
                str(Path(settings.ROOT_PATH).resolve()),
                readonly=False,
            ),
        },
        
        # 服务器设置
        "host": "0.0.0.0",
        "port": 8080,  # 这里仅用于配置，实际由 ASGI 服务器控制
        
        # 认证配置
        "http_authenticator": {
            "domain_controller": FluxFileDomainController(),
            "accept_basic": True,
            "accept_digest": False,
            "default_to_digest": False,
        },
        
        # 简化认证（开发模式）
        "simple_dc": {
            "user_mapping": {
                "*": {
                    "admin": {
                        "password": "admin",
                        "roles": ["admin"],
                    },
                    "user": {
                        "password": "user",
                        "roles": ["user"],
                    },
                } if settings.DEBUG else {}
            },
        },
        
        # 日志和调试
        "verbose": 1 if settings.DEBUG else 0,
        "logging": {
            "enable_loggers": ["wsgidav"],
        },
        
        # 锁管理器
        "lock_storage": True,
        
        # 属性管理器
        "property_manager": True,
        
        # 中间件
        "middleware_stack": [
            # 添加自定义中间件
            # FluxFileMiddleware,  # 如果需要自定义处理
        ],
        
        # 杂项
        "hotfixes": {
            # "re_encode_path_info": True,
        },
    }
    
    # 创建应用
    app = WsgiDAVApp(config)
    
    # 包装中间件
    app = FluxFileMiddleware(app)
    
    logger.info(f"WebDAV 服务已创建，根目录: {settings.ROOT_PATH}")
    
    return app


# ============================================================================
# 挂载到 FastAPI
# ============================================================================

def mount_webdav(app):
    """
    将 WebDAV 应用挂载到 FastAPI
    
    使用 WSGIMiddleware 将 WSGI 应用适配为 ASGI。
    
    挂载路径：/webdav
    
    使用示例：
        from fastapi import FastAPI
        from app.integrations.webdav import mount_webdav
        
        app = FastAPI()
        mount_webdav(app)
    """
    try:
        from starlette.middleware.wsgi import WSGIMiddleware
    except ImportError:
        logger.error("Starlette 缺少 WSGIMiddleware，请确保安装了完整版本")
        return
    
    # 创建 WebDAV 应用
    webdav_app = create_webdav_app()
    if webdav_app is None:
        logger.warning("WebDAV 应用创建失败，功能已禁用")
        return
    
    # 使用 WSGIMiddleware 包装并挂载
    try:
        app.mount("/webdav", WSGIMiddleware(webdav_app))
        logger.info("WebDAV 已挂载到 /webdav")
    except Exception as e:
        logger.error(f"WebDAV 挂载失败: {e}")


# ============================================================================
# 带认证上下文的 WebDAV
# ============================================================================

class AuthenticatedWebDAVMiddleware:
    """
    认证感知的 WebDAV 中间件
    
    从 FastAPI 的认证上下文中提取用户信息，
    传递给 WsgiDAV 进行权限控制。
    """
    
    def __init__(
        self,
        webdav_app: Callable,
        auth_header_name: str = "Authorization",
    ):
        self.webdav_app = webdav_app
        self.auth_header_name = auth_header_name
    
    def __call__(
        self,
        environ: Dict[str, Any],
        start_response: Callable,
    ):
        # 尝试从 JWT Token 解析用户信息
        auth_header = environ.get(f"HTTP_{self.auth_header_name.upper()}", "")
        
        if auth_header.startswith("Bearer "):
            token = auth_header[7:]
            user_info = self._decode_jwt(token)
            
            if user_info:
                # 将用户信息注入到环境中
                environ["wsgidav.user"] = user_info.get("username")
                environ["wsgidav.roles"] = user_info.get("roles", [])
        
        return self.webdav_app(environ, start_response)
    
    def _decode_jwt(self, token: str) -> Optional[Dict[str, Any]]:
        """解码 JWT Token"""
        try:
            from jose import jwt
            
            payload = jwt.decode(
                token,
                settings.SECRET_KEY,
                algorithms=["HS256"],
            )
            return payload
        except Exception as e:
            logger.debug(f"JWT 解码失败: {e}")
            return None


def mount_webdav_with_auth(app):
    """
    挂载带认证的 WebDAV
    
    此版本会尝试从 FastAPI 的 JWT Token 解析用户信息。
    """
    try:
        from starlette.middleware.wsgi import WSGIMiddleware
    except ImportError:
        logger.error("Starlette 缺少 WSGIMiddleware")
        return
    
    webdav_app = create_webdav_app()
    if webdav_app is None:
        return
    
    # 添加认证中间件
    webdav_app = AuthenticatedWebDAVMiddleware(webdav_app)
    
    # 挂载
    app.mount("/webdav", WSGIMiddleware(webdav_app))
    logger.info("WebDAV（带认证）已挂载到 /webdav")
